Migration to Next.js
Migrate the entire project to Next.js.
Fix linting for some files.
