+++ .prettierignore
@@ -1,0 +1,5 @@
+node_modules
+.next
+yarn.lock
+package-lock.json
+public
