--- src/index.tsx
@@ -1,20 +1,0 @@
-// pre-generated
-import React from 'react';
-import ReactDOM from 'react-dom';
-import * as serviceWorker from './serviceWorker';
-
-// bootstrap
-import "bootstrap/dist/css/bootstrap.min.css"
-
-// local styles
-import "./index.css"
-import "./array-element.css"
-
-import SortViz from './SortViz'
-
-ReactDOM.render(<SortViz />, document.getElementById('root'));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
